Eslint

+++ src/Path/Dashboard.jsx
@@ -40,7 +40,7 @@ const Dashboard = () => {
 		});
 		if (window.location.pathname == "/Dashboard") {
 			setViewAllListings(false);
-		} else  {
+		} else {
 			setViewAllListings(true);
 		}
 		document.title = "Dashboard";
@@ -49,11 +49,10 @@ const Dashboard = () => {
 	useEffect(() => {
 		if (pageNo == 1) {
 			fetchListings();
-		}
-		else {
+		} else {
 			// setPageNo(1);
 			fetchListings();
-			}
+		}
 	}, [selectedStatus, viewAllListings, pageNo]);
 
 	let navigate = useNavigate();
@@ -73,12 +72,12 @@ const Dashboard = () => {
 	function fetchListings() {
 		if (viewAllListings == true) {
 			getListings({ statusId: selectedStatus, pageNo }).then((response) => {
-				setListings((response.data.data));
+				setListings(response.data.data);
 			});
 		}
-		if (viewAllListings == false){
+		if (viewAllListings == false) {
 			getUserListings({ statusId: selectedStatus, pageNo }).then((response) => {
-				setListings((response.data.data));
+				setListings(response.data.data);
 			});
 		}
 	}
@@ -112,30 +111,34 @@ const Dashboard = () => {
 		);
 	}
 
-		const [showConfirmationModal, setShowConfirmationModal] = useState({
-			visible: false,
-			listing: null,
-			onConfirm: () => {},
-			onCancel: () => {}
-		});
+	const [showConfirmationModal, setShowConfirmationModal] = useState({
+		visible: false,
+		listing: null,
+		onConfirm: () => {},
+		onCancel: () => {},
+	});
 
-		function handleDelete(listing) {
-			deleteListing(listing.cityId, listing.id)
-			  .then((res) => {
-				setListings(listings.filter((l) => l.cityId !== listing.cityId || l.id !== listing.id));
+	function handleDelete(listing) {
+		deleteListing(listing.cityId, listing.id)
+			.then((res) => {
+				setListings(
+					listings.filter(
+						(l) => l.cityId !== listing.cityId || l.id !== listing.id
+					)
+				);
 				setShowConfirmationModal({ visible: false }); // hide the confirmation modal
-			  })
-			  .catch((error) => console.log(error));
-		  }
+			})
+			.catch((error) => console.log(error));
+	}
 
-		function deleteListingOnClick(listing) {
-			setShowConfirmationModal({
+	function deleteListingOnClick(listing) {
+		setShowConfirmationModal({
 			visible: true,
 			listing,
 			onConfirm: () => handleDelete(listing),
 			onCancel: () => setShowConfirmationModal({ visible: false }),
-			});
-		}
+		});
+	}
 
 	function goToEventDetailsPage(listing) {
 		navigateTo(
@@ -147,8 +150,7 @@ const Dashboard = () => {
 
 	return (
 		<section className="bg-slate-600 body-font relative">
-			<SideBar
-			/>
+			<SideBar />
 
 			<div class="container px-0 sm:px-0 py-0 w-full fixed top-0 z-10 lg:px-5 lg:w-auto lg:relative">
 				<Popover className="relative bg-black mr-0 ml-0 px-10 lg:rounded-lg h-16">
@@ -159,25 +161,29 @@ const Dashboard = () => {
 									<div
 										class="text-gray-300 hover:bg-gray-700 hover:text-white rounded-md p-4 text-sm font-bold cursor-pointer"
 										onClick={() => setSelectedStatus(null)}
-										style={{ fontFamily: 'Poppins, sans-serif' }}>
+										style={{ fontFamily: "Poppins, sans-serif" }}
+									>
 										{t("allListings")}
 									</div>
 									<div
 										class="text-gray-300 hover:bg-gray-700 hover:text-white rounded-md p-4 text-sm font-bold cursor-pointer"
 										onClick={() => setSelectedStatus(statusByName.Active)}
-										style={{ fontFamily: 'Poppins, sans-serif' }}>
+										style={{ fontFamily: "Poppins, sans-serif" }}
+									>
 										{t("active")}
 									</div>
 									<div
 										class="text-gray-300 hover:bg-gray-700 hover:text-white rounded-md p-4 text-sm font-bold cursor-pointer"
 										onClick={() => setSelectedStatus(statusByName.Pending)}
-										style={{ fontFamily: 'Poppins, sans-serif' }}>
+										style={{ fontFamily: "Poppins, sans-serif" }}
+									>
 										{t("pending")}
 									</div>
 									<div
 										class="text-gray-300 hover:bg-gray-700 hover:text-white rounded-md p-4 text-sm font-bold cursor-pointer"
 										onClick={() => setSelectedStatus(statusByName.Inactive)}
-										style={{ fontFamily: 'Poppins, sans-serif' }}>
+										style={{ fontFamily: "Poppins, sans-serif" }}
+									>
 										{t("inactive")}
 									</div>
 								</div>
@@ -223,13 +229,15 @@ const Dashboard = () => {
 												<div
 													class="text-gray-300 hover:bg-gray-700 hover:text-white rounded-md p-4 text-sm font-bold cursor-pointer"
 													onClick={() => setSelectedStatus(null)}
-													style={{ fontFamily: 'Poppins, sans-serif' }}>
+													style={{ fontFamily: "Poppins, sans-serif" }}
+												>
 													{t("allListings")}
 												</div>
 												<div
 													class="text-gray-300 hover:bg-gray-700 hover:text-white rounded-md p-4 text-sm font-bold cursor-pointer"
 													onClick={() => setSelectedStatus(statusByName.Active)}
-													style={{ fontFamily: 'Poppins, sans-serif' }}>
+													style={{ fontFamily: "Poppins, sans-serif" }}
+												>
 													{t("active")}
 												</div>
 												<div
@@ -237,7 +245,8 @@ const Dashboard = () => {
 													onClick={() =>
 														setSelectedStatus(statusByName.Pending)
 													}
-													style={{ fontFamily: 'Poppins, sans-serif' }}>
+													style={{ fontFamily: "Poppins, sans-serif" }}
+												>
 													{t("pending")}
 												</div>
 												<div
@@ -245,7 +254,8 @@ const Dashboard = () => {
 													onClick={() =>
 														setSelectedStatus(statusByName.Inactive)
 													}
-													style={{ fontFamily: 'Poppins, sans-serif' }}>
+													style={{ fontFamily: "Poppins, sans-serif" }}
+												>
 													{t("inactive")}
 												</div>
 											</div>
@@ -266,24 +276,48 @@ const Dashboard = () => {
 					<table class="w-full text-sm text-left lg:mt-[2rem] mt-[2rem] text-gray-500 dark:text-gray-400 p-6 space-y-10 rounded-lg">
 						<thead class="text-xs text-gray-700 uppercase bg-gray-50 dark:bg-gray-50 dark:text-gray-700">
 							<tr>
-								<th scope="col" class="px-6 sm:px-6 py-3" style={{ fontFamily: 'Poppins, sans-serif' }}>
+								<th
+									scope="col"
+									class="px-6 sm:px-6 py-3"
+									style={{ fontFamily: "Poppins, sans-serif" }}
+								>
 									{t("listings")}
 								</th>
-								<th scope="col" class="px-6 sm:px-3 py-3 hidden lg:table-cell" style={{ fontFamily: 'Poppins, sans-serif' }}>
+								<th
+									scope="col"
+									class="px-6 sm:px-3 py-3 hidden lg:table-cell"
+									style={{ fontFamily: "Poppins, sans-serif" }}
+								>
 									{t("category")}
 								</th>
-								<th scope="col" class="px-6 py-3 hidden lg:table-cell" style={{ fontFamily: 'Poppins, sans-serif' }}>
+								<th
+									scope="col"
+									class="px-6 py-3 hidden lg:table-cell"
+									style={{ fontFamily: "Poppins, sans-serif" }}
+								>
 									{t("date_of_creation")}
 								</th>
-								<th scope="col" class="px-6 py-3" style={{ fontFamily: 'Poppins, sans-serif' }}>
+								<th
+									scope="col"
+									class="px-6 py-3"
+									style={{ fontFamily: "Poppins, sans-serif" }}
+								>
 									{t("action")}
 								</th>
 								{viewAllListings && (
-									<th scope="col" class="px-6 py-3" style={{ fontFamily: 'Poppins, sans-serif' }}>
+									<th
+										scope="col"
+										class="px-6 py-3"
+										style={{ fontFamily: "Poppins, sans-serif" }}
+									>
 										{t("username")}
 									</th>
 								)}
-								<th scope="col" class="px-6 py-3 text-center" style={{ fontFamily: 'Poppins, sans-serif' }}>
+								<th
+									scope="col"
+									class="px-6 py-3 text-center"
+									style={{ fontFamily: "Poppins, sans-serif" }}
+								>
 									{t("status")}
 								</th>
 							</tr>
@@ -307,71 +341,116 @@ const Dashboard = () => {
 												alt="avatar"
 											/>
 											<div class="pl-0 sm:pl-3 overflow-hidden max-w-[20rem]">
-												<div class="font-normal text-gray-500 truncate" style={{ fontFamily: 'Poppins, sans-serif' }}>
+												<div
+													class="font-normal text-gray-500 truncate"
+													style={{ fontFamily: "Poppins, sans-serif" }}
+												>
 													{listing.title}
 												</div>
 											</div>
 										</th>
-										<td class="px-6 py-4 hidden lg:table-cell" style={{ fontFamily: 'Poppins, sans-serif' }}>
+										<td
+											class="px-6 py-4 hidden lg:table-cell"
+											style={{ fontFamily: "Poppins, sans-serif" }}
+										>
 											{categoryById[listing.categoryId]}
 										</td>
-										<td class="px-6 py-4 hidden lg:table-cell" style={{ fontFamily: 'Poppins, sans-serif' }}>
+										<td
+											class="px-6 py-4 hidden lg:table-cell"
+											style={{ fontFamily: "Poppins, sans-serif" }}
+										>
 											{new Date(listing.createdAt).toLocaleString("de")}
 										</td>
 										<td class="px-6 py-4">
 											<a
 												class="font-medium text-blue-600 dark:text-blue-500 hover:underline cursor-pointer pr-2"
 												onClick={() => goToEditListingsPage(listing)}
-												style={{ fontFamily: 'Poppins, sans-serif' }}>
+												style={{ fontFamily: "Poppins, sans-serif" }}
+											>
 												{t("edit")}
 											</a>
 											<a
-											className="font-medium text-blue-600 dark:text-blue-500 hover:underline cursor-pointer"
-											onClick={() => deleteListingOnClick(listing)}
-											style={{ fontFamily: 'Poppins, sans-serif' }}>
-											{t("delete")}
+												className="font-medium text-blue-600 dark:text-blue-500 hover:underline cursor-pointer"
+												onClick={() => deleteListingOnClick(listing)}
+												style={{ fontFamily: "Poppins, sans-serif" }}
+											>
+												{t("delete")}
 											</a>
 											{showConfirmationModal.visible && (
 												<div className="fixed z-10 inset-0 overflow-y-auto">
-												<div className="flex items-end justify-center min-h-screen pt-4 px-4 pb-20 text-center sm:block sm:p-0">
-													<div className="fixed inset-0 transition-opacity" aria-hidden="true">
-													<div className="absolute inset-0 bg-gray-500 opacity-75"></div>
-													</div>
-													<span className="hidden sm:inline-block sm:align-middle sm:h-screen" aria-hidden="true">&#8203;</span>
-													<div className="inline-block align-bottom bg-white rounded-lg text-left overflow-hidden shadow-xl transform transition-all sm:my-8 sm:align-middle sm:max-w-lg sm:w-full">
-													<div className="bg-white px-4 pt-5 pb-4 sm:p-6 sm:pb-4">
-														<div className="sm:flex sm:items-start">
-														<div className="mx-auto flex-shrink-0 flex items-center justify-center h-12 w-12 rounded-full bg-red-100 sm:mx-0 sm:h-10 sm:w-10">
-															<svg className="h-6 w-6 text-red-600" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke="currentColor" aria-hidden="true">
-															<path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M6 18L18 6M6 6l12 12" />
-															</svg>
+													<div className="flex items-end justify-center min-h-screen pt-4 px-4 pb-20 text-center sm:block sm:p-0">
+														<div
+															className="fixed inset-0 transition-opacity"
+															aria-hidden="true"
+														>
+															<div className="absolute inset-0 bg-gray-500 opacity-75"></div>
 														</div>
-														<div className="mt-3 text-center sm:mt-0 sm:ml-4 sm:text-left">
-															<h3 className="text-lg leading-6 font-medium text-gray-900">{t("areyousure")}</h3>
-															<div className="mt-2">
-															<p className="text-sm text-gray-500">{t("doyoureallywanttodeleteListing")}</p>
+														<span
+															className="hidden sm:inline-block sm:align-middle sm:h-screen"
+															aria-hidden="true"
+														>
+															&#8203;
+														</span>
+														<div className="inline-block align-bottom bg-white rounded-lg text-left overflow-hidden shadow-xl transform transition-all sm:my-8 sm:align-middle sm:max-w-lg sm:w-full">
+															<div className="bg-white px-4 pt-5 pb-4 sm:p-6 sm:pb-4">
+																<div className="sm:flex sm:items-start">
+																	<div className="mx-auto flex-shrink-0 flex items-center justify-center h-12 w-12 rounded-full bg-red-100 sm:mx-0 sm:h-10 sm:w-10">
+																		<svg
+																			className="h-6 w-6 text-red-600"
+																			xmlns="http://www.w3.org/2000/svg"
+																			fill="none"
+																			viewBox="0 0 24 24"
+																			stroke="currentColor"
+																			aria-hidden="true"
+																		>
+																			<path
+																				strokeLinecap="round"
+																				strokeLinejoin="round"
+																				strokeWidth="2"
+																				d="M6 18L18 6M6 6l12 12"
+																			/>
+																		</svg>
+																	</div>
+																	<div className="mt-3 text-center sm:mt-0 sm:ml-4 sm:text-left">
+																		<h3 className="text-lg leading-6 font-medium text-gray-900">
+																			{t("areyousure")}
+																		</h3>
+																		<div className="mt-2">
+																			<p className="text-sm text-gray-500">
+																				{t("doyoureallywanttodeleteListing")}
+																			</p>
+																		</div>
+																	</div>
+																</div>
+															</div>
+															<div className="bg-gray-50 px-4 py-3 sm:px-6 sm:flex sm:flex-row-reverse">
+																<button
+																	onClick={showConfirmationModal.onConfirm}
+																	type="button"
+																	className="w-full inline-flex justify-center rounded-md border border-transparent shadow-sm px-4 py-2 bg-red-600 text-base font-medium text-white hover:bg-red-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-red-500 sm:ml-3 sm:w-auto sm:text-sm"
+																>
+																	{t("delete")}
+																</button>
+
+																<button
+																	onClick={showConfirmationModal.onCancel}
+																	type="button"
+																	className="mt-3 w-full inline-flex justify-center rounded-md border border-gray-300 shadow-sm px-4 py-2 bg-white text-base font-medium text-gray-700 hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500 sm:mt-0 sm:w-auto sm:text-sm"
+																>
+																	{t("cancel")}
+																</button>
 															</div>
 														</div>
-														</div>
 													</div>
-													<div className="bg-gray-50 px-4 py-3 sm:px-6 sm:flex sm:flex-row-reverse">
-													<button onClick={showConfirmationModal.onConfirm} type="button" className="w-full inline-flex justify-center rounded-md border border-transparent shadow-sm px-4 py-2 bg-red-600 text-base font-medium text-white hover:bg-red-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-red-500 sm:ml-3 sm:w-auto sm:text-sm">
-														{t("delete")}
-													</button>
-
-													<button onClick={showConfirmationModal.onCancel} type="button" className="mt-3 w-full inline-flex justify-center rounded-md border border-gray-300 shadow-sm px-4 py-2 bg-white text-base font-medium text-gray-700 hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500 sm:mt-0 sm:w-auto sm:text-sm">
-														{t("cancel")}
-													</button>
-
-													</div>
-													</div>
-												</div>
 												</div>
 											)}
 										</td>
 										{viewAllListings && (
 											<td class="px-6 py-4">
-												<a class="font-medium text-blue-600 dark:text-blue-500 hover:underline cursor-pointer" style={{ fontFamily: 'Poppins, sans-serif' }}>
+												<a
+													class="font-medium text-blue-600 dark:text-blue-500 hover:underline cursor-pointer"
+													style={{ fontFamily: "Poppins, sans-serif" }}
+												>
 													{listing.username}
 												</a>
 											</td>
@@ -389,7 +468,8 @@ const Dashboard = () => {
 															handleChangeInStatus(e.target.value, listing)
 														}
 														value={listing.statusId}
-														style={{ fontFamily: 'Poppins, sans-serif' }}>
+														style={{ fontFamily: "Poppins, sans-serif" }}
+													>
 														{Object.keys(status).map((state) => {
 															return (
 																<>
@@ -401,7 +481,9 @@ const Dashboard = () => {
 														})}
 													</Select>
 												) : (
-													<h1 style={{ fontFamily: 'Poppins, sans-serif' }}>{status[listing.statusId]}</h1>
+													<h1 style={{ fontFamily: "Poppins, sans-serif" }}>
+														{status[listing.statusId]}
+													</h1>
 												)}
 											</div>
 										</td>
@@ -416,22 +498,27 @@ const Dashboard = () => {
 						<span
 							className="text-md px-3 hover:bg-gray-800 cursor-pointer rounded-lg"
 							onClick={() => setPageNo(pageNo - 1)}
-							style={{ fontFamily: 'Poppins, sans-serif' }}>
+							style={{ fontFamily: "Poppins, sans-serif" }}
+						>
 							{"<"}{" "}
 						</span>
 					) : (
 						<span />
 					)}
-					<span className="text-lg px-3" style={{ fontFamily: 'Poppins, sans-serif' }}>
+					<span
+						className="text-lg px-3"
+						style={{ fontFamily: "Poppins, sans-serif" }}
+					>
 						{t("page")} {pageNo}
 					</span>
 
 					{listings.length >= 9 && (
 						<span
-						className="text-lg px-3 hover:bg-gray-800 cursor-pointer rounded-lg"
-						onClick={() => setPageNo(pageNo + 1)}
-						style={{ fontFamily: 'Poppins, sans-serif' }}>
-						{">"}
+							className="text-lg px-3 hover:bg-gray-800 cursor-pointer rounded-lg"
+							onClick={() => setPageNo(pageNo + 1)}
+							style={{ fontFamily: "Poppins, sans-serif" }}
+						>
+							{">"}
 						</span>
 					)}
 				</div>
